Package fragment moved into view. New package activity created. Moved some classes into package activity
New Classes created: Suitcase.java, CarrierActivity.java, SuitcaseViewHolder.java.

New layouts created: item_suitcase.xml and activity_carrier.xml.

Show suitcase data to carriers feature added on CarrierActivity.java
