+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".TrackingActivity" />
-        <activity android:name=".view.SlideActivity" />
-        <activity android:name=".view.TestingActivity" />
-        <activity android:name=".view.BottomNavigationViewActivity" />
-        <activity android:name=".view.SelectLoginActivity" />
-        <activity android:name=".view.RegisterActivity" />
-        <activity android:name=".view.LoginActivity" />
+        <activity android:name=".view.activity.CarrierActivity"></activity>
+        <activity android:name=".view.activity.TrackingActivity" />
+        <activity android:name=".view.activity.SlideActivity" />
+        <activity android:name=".view.activity.TestingActivity" />
+        <activity android:name=".view.activity.BottomNavigationViewActivity" />
+        <activity android:name=".view.activity.SelectLoginActivity" />
+        <activity android:name=".view.activity.RegisterActivity" />
+        <activity android:name=".view.activity.LoginActivity" />
         <activity
             android:name=".MainActivity"
             android:theme="@style/SplashTheme">
