--- app/src/main/java/com/example/koffer/RecyclerViewAdapter.java
@@ -1,4 +1,0 @@
-package com.example.koffer;
-
-public class RecyclerViewAdapter {
-}
