+++ app/src/main/java/com/example/koffer/model/Suitcase.java
@@ -1,0 +1,19 @@
+package com.example.koffer.model;
+
+public class Suitcase {
+    public String name;
+    public String email;
+    public String quantity;
+    public String kg;
+
+    public Suitcase() {
+
+    }
+
+    public Suitcase(String name, String email, String quantity, String kg) {
+        this.name = name;
+        this.email = email;
+        this.quantity = quantity;
+        this.kg = kg;
+    }
+}
