+++ app/src/main/java/com/example/koffer/view/activity/CarrierActivity.java
@@ -1,0 +1,67 @@
+package com.example.koffer.view.activity;
+
+import android.support.annotation.NonNull;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.example.koffer.R;
+import com.example.koffer.model.Suitcase;
+import com.example.koffer.view.SuitcaseViewHolder;
+import com.firebase.ui.database.FirebaseRecyclerAdapter;
+import com.firebase.ui.database.FirebaseRecyclerOptions;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.Query;
+
+public class CarrierActivity extends AppCompatActivity {
+
+    public static final String SUITCASE_REFERENCE = "suitcase";
+
+    public DatabaseReference mReference;
+    public FirebaseUser mUser;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_carrier);
+
+        mReference = FirebaseDatabase.getInstance().getReference();
+        mUser = FirebaseAuth.getInstance().getCurrentUser();
+
+        FirebaseRecyclerOptions options = new FirebaseRecyclerOptions.Builder<Suitcase>()
+                .setIndexedQuery(setQuery(), mReference.child(SUITCASE_REFERENCE), Suitcase.class)
+                .setLifecycleOwner(this)
+                .build();
+
+        RecyclerView recyclerView = findViewById(R.id.suitcase_list);
+        recyclerView.setLayoutManager(new LinearLayoutManager(this));
+        recyclerView.setAdapter(new FirebaseRecyclerAdapter<Suitcase, SuitcaseViewHolder>(options) {
+            @NonNull
+            @Override
+            public SuitcaseViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int i) {
+                View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.item_suitecase, parent, false);
+
+                return new SuitcaseViewHolder(view);
+            }
+
+            @Override
+            protected void onBindViewHolder(@NonNull SuitcaseViewHolder holder, int position, @NonNull Suitcase suitcase) {
+                holder.userName.setText(suitcase.name);
+                holder.userEmail.setText(suitcase.email);
+                holder.suitcaseQuantity.setText(suitcase.quantity);
+                holder.suitcaseKG.setText(suitcase.kg);
+            }
+        });
+    }
+
+    Query setQuery(){
+        return  mReference.child(SUITCASE_REFERENCE).limitToFirst(100);
+    }
+}
