+++ app/src/main/java/com/example/koffer/view/activity/SelectLoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.koffer.view;
+package com.example.koffer.view.activity;
 
 import android.content.Context;
 import android.content.Intent;
@@ -9,7 +9,6 @@ import android.view.View;
 import android.widget.Button;
 
 import com.example.koffer.R;
-import com.example.koffer.TrackingActivity;
 
 public class SelectLoginActivity extends AppCompatActivity {
