+++ app/src/main/java/com/example/koffer/view/activity/TrackingActivity.java
@@ -1,4 +1,4 @@
-package com.example.koffer;
+package com.example.koffer.view.activity;
 
 import android.Manifest;
 import android.content.pm.PackageManager;
@@ -8,6 +8,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 
+import com.example.koffer.R;
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.tasks.OnSuccessListener;
