+++ app/src/main/java/com/example/koffer/view/fragment/MapFragment.java
@@ -1,4 +1,4 @@
-package com.example.koffer.fragment;
+package com.example.koffer.view.fragment;
 
 
 import android.os.Bundle;
