+++ app/src/main/java/com/example/koffer/view/fragment/ServiceFragment.java
@@ -1,4 +1,4 @@
-package com.example.koffer.fragment;
+package com.example.koffer.view.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
