+++ app/src/main/res/layout/activity_bottom_navigation_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.BottomNavigationViewActivity">
+    tools:context=".view.activity.BottomNavigationViewActivity">
 
     <FrameLayout
         android:id="@+id/fragment_container"
