+++ app/src/main/res/layout/activity_carrier.xml
@@ -4,6 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.activity.TrackingActivity">
+    tools:context=".view.activity.CarrierActivity">
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/suitcase_list"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:scrollbars="vertical" />
 
 </android.support.constraint.ConstraintLayout>
