+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.RegisterActivity"
+    tools:context=".view.activity.RegisterActivity"
     android:orientation="vertical"
     android:background="@color/colorPrimary"
     android:padding="25dp">
