+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/holo_blue_bright"
-    tools:context=".fragment.HomeFragment">
+    tools:context=".view.fragment.HomeFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
