+++ app/src/main/res/layout/fragment_service.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@android:color/holo_orange_dark"
-    tools:context=".fragment.ServiceFragment">
+    tools:context=".view.fragment.ServiceFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
