+++ app/src/main/res/values/strings.xml
@@ -17,11 +17,18 @@
     <string name="action_settings">Settings</string>
 
     <!-- Toast messages -->
-    <string name="logging_in">Iniciando sesión...</string>
+    <string name="logging_in">Iniciando sesión…</string>
     <string name="logging_in_error">Error al inicar sesión, prueba de nuevo mas tarde</string>
     <string name="welcome">Bienvenido:</string>
     <string name="success">Sesión inicada correctamente</string>
     <string name="empty_fields">Se deben rellenar todos los campos</string>
     <string name="user_not_registered">¡Este usuario no está registrado!</string>
+
+    <!-- Suitcase -->
+    <string name="user_name">User Name</string>
+    <string name="user_email">User Email</string>
+    <string name="suitcase_quantity">Suitcase Quantity</string>
+    <string name="suitcase_kg">Suitcase KG</string>
+
 
 </resources>
